Refactor TitleField Move off Review-page title-manipulation method to helpers
